requirements.txt
moviepy>=2.0.0
edge-tts>=6.1.0
openai-whisper>=20230918
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0
pydub>=0.25.1
audioop-lts>=0.2.1
rich>=13.0.0
